feat: add repository for task_template
