+++ services/task-svc/internal/task-template/models/repository.go
@@ -1,0 +1,15 @@
+package models
+
+import (
+	"gorm.io/gorm"
+)
+
+type TemplateRepository struct {
+	db *gorm.DB
+}
+
+func NewTemplateRepositoryWithDB(db *gorm.DB) *TemplateRepository {
+	return &TemplateRepository{
+		db: db,
+	}
+}
