+++ services/task-svc/internal/task-template/task_template.go
@@ -11,28 +11,8 @@ import (
 	"hwgorm"
 	"hwutil"
 	pbhelpers "proto_helpers/task_svc/v1"
+	templatemodels "task-svc/internal/task-template/models"
 )
-
-type Base struct {
-	Name        string `gorm:"column:name"`
-	Description string `gorm:"column:description"`
-}
-
-type TaskTemplate struct {
-	Base
-	ID             uuid.UUID             `gorm:"column:id"`
-	OrganizationID uuid.UUID             `gorm:"column:organization_id"`
-	Public         bool                  `gorm:"column:is_public"`
-	UserID         uuid.UUID             `gorm:"column:user_id"`
-	WardID         *uuid.UUID            `gorm:"column:ward_id;default:NULL"`
-	SubTasks       []TaskTemplateSubtask `gorm:"foreignKey:TaskTemplateID"`
-}
-
-type TaskTemplateSubtask struct {
-	ID             uuid.UUID `gorm:"column:id"`
-	TaskTemplateID uuid.UUID `gorm:"column:task_template_id"`
-	Name           string    `gorm:"column:name"`
-}
 
 type ServiceServer struct {
 	pb.UnimplementedTaskTemplateServiceServer
@@ -61,8 +41,8 @@ func (ServiceServer) CreateTaskTemplate(ctx context.Context, req *pb.CreateTaskT
 		return nil, err
 	}
 
-	taskTemplate := TaskTemplate{
-		Base:           Base{Name: req.Name, Description: req.Description},
+	taskTemplate := templatemodels.TaskTemplate{
+		Base:           templatemodels.Base{Name: req.Name, Description: req.Description},
 		OrganizationID: organizationID,
 		Public:         req.Public,
 		UserID:         userID,
@@ -74,10 +54,10 @@ func (ServiceServer) CreateTaskTemplate(ctx context.Context, req *pb.CreateTaskT
 		return nil, status.Error(codes.Internal, err.Error())
 	}
 
-	var subtasks []TaskTemplateSubtask
+	var subtasks []templatemodels.TaskTemplateSubtask
 	if req.Subtasks != nil {
-		subtasks = hwutil.Map(req.Subtasks, func(subtask *pb.CreateTaskTemplateRequest_SubTask) TaskTemplateSubtask {
-			return TaskTemplateSubtask{Name: subtask.Name, TaskTemplateID: taskTemplate.ID}
+		subtasks = hwutil.Map(req.Subtasks, func(subtask *pb.CreateTaskTemplateRequest_SubTask) templatemodels.TaskTemplateSubtask {
+			return templatemodels.TaskTemplateSubtask{Name: subtask.Name, TaskTemplateID: taskTemplate.ID}
 		})
 	}
 
@@ -103,7 +83,7 @@ func (ServiceServer) GetAllTaskTemplates(ctx context.Context, _ *pb.GetAllTaskTe
 		return nil, status.Error(codes.InvalidArgument, err.Error())
 	}
 
-	var taskTemplates []TaskTemplate
+	var taskTemplates []templatemodels.TaskTemplate
 
 	if err := db.Preload("SubTasks").Where("organization_id = ?", organizationID).Where("is_public = ?", true).Find(&taskTemplates).Error; err != nil {
 		if hwgorm.IsOurFault(err) {
@@ -113,8 +93,8 @@ func (ServiceServer) GetAllTaskTemplates(ctx context.Context, _ *pb.GetAllTaskTe
 		}
 	}
 
-	var mappedTaskTemplates = hwutil.Map(taskTemplates, func(taskTemplate TaskTemplate) *pb.GetAllTaskTemplatesResponse_TaskTemplate {
-		var mappedSubtasks = hwutil.Map(taskTemplate.SubTasks, func(subtask TaskTemplateSubtask) *pb.GetAllTaskTemplatesResponse_TaskTemplate_SubTask {
+	var mappedTaskTemplates = hwutil.Map(taskTemplates, func(taskTemplate templatemodels.TaskTemplate) *pb.GetAllTaskTemplatesResponse_TaskTemplate {
+		var mappedSubtasks = hwutil.Map(taskTemplate.SubTasks, func(subtask templatemodels.TaskTemplateSubtask) *pb.GetAllTaskTemplatesResponse_TaskTemplate_SubTask {
 			return &pb.GetAllTaskTemplatesResponse_TaskTemplate_SubTask{
 				Id:             subtask.ID.String(),
 				TaskTemplateId: subtask.TaskTemplateID.String(),
@@ -147,7 +127,7 @@ func (ServiceServer) DeleteTaskTemplate(ctx context.Context, req *pb.DeleteTaskT
 		return nil, status.Error(codes.InvalidArgument, err.Error())
 	}
 
-	taskTemplate := TaskTemplate{ID: id}
+	taskTemplate := templatemodels.TaskTemplate{ID: id}
 
 	if err := db.Delete(&taskTemplate).Error; err != nil {
 		return nil, status.Error(codes.Internal, err.Error())
@@ -171,7 +151,7 @@ func (ServiceServer) DeleteTaskTemplateSubTask(ctx context.Context, req *pb.Dele
 		return nil, status.Error(codes.InvalidArgument, err.Error())
 	}
 
-	taskTemplateSubtask := TaskTemplateSubtask{ID: id}
+	taskTemplateSubtask := templatemodels.TaskTemplateSubtask{ID: id}
 
 	if err := db.Delete(&taskTemplateSubtask).Error; err != nil {
 		return nil, status.Error(codes.Internal, err.Error())
@@ -195,7 +175,7 @@ func (ServiceServer) UpdateTaskTemplate(ctx context.Context, req *pb.UpdateTaskT
 		return nil, status.Error(codes.InvalidArgument, err.Error())
 	}
 
-	taskTemplate := TaskTemplate{ID: id}
+	taskTemplate := templatemodels.TaskTemplate{ID: id}
 	updates := pbhelpers.UpdatesMapForUpdateTaskTemplateRequest(req)
 
 	if err := db.Model(&taskTemplate).Updates(updates).Error; err != nil {
@@ -215,7 +195,7 @@ func (ServiceServer) UpdateTaskTemplateSubtask(ctx context.Context, req *pb.Upda
 		return nil, status.Error(codes.InvalidArgument, err.Error())
 	}
 
-	taskTemplate := TaskTemplateSubtask{ID: id}
+	taskTemplate := templatemodels.TaskTemplateSubtask{ID: id}
 	updates := pbhelpers.UpdatesMapForUpdateTaskTemplateSubTaskRequest(req)
 
 	if err := db.Model(&taskTemplate).Updates(updates).Error; err != nil {
@@ -234,7 +214,7 @@ func (ServiceServer) CreateTaskTemplateSubTask(ctx context.Context, req *pb.Crea
 		return nil, err
 	}
 
-	taskTemplateSubtask := TaskTemplateSubtask{
+	taskTemplateSubtask := templatemodels.TaskTemplateSubtask{
 		TaskTemplateID: taskTemplateID,
 		Name:           req.Name,
 	}
@@ -264,7 +244,7 @@ func (ServiceServer) GetAllTaskTemplatesByUser(ctx context.Context, req *pb.GetA
 		return nil, status.Error(codes.InvalidArgument, err.Error())
 	}
 
-	var taskTemplates []TaskTemplate
+	var taskTemplates []templatemodels.TaskTemplate
 
 	if err := db.Preload("SubTasks").Where("user_id = ?", userId).Find(&taskTemplates).Error; err != nil {
 		if hwgorm.IsOurFault(err) {
@@ -274,8 +254,8 @@ func (ServiceServer) GetAllTaskTemplatesByUser(ctx context.Context, req *pb.GetA
 		}
 	}
 
-	var mappedTaskTemplates = hwutil.Map(taskTemplates, func(taskTemplate TaskTemplate) *pb.GetAllTaskTemplatesByUserResponse_TaskTemplate {
-		var mappedSubtasks = hwutil.Map(taskTemplate.SubTasks, func(subtask TaskTemplateSubtask) *pb.GetAllTaskTemplatesByUserResponse_TaskTemplate_SubTask {
+	var mappedTaskTemplates = hwutil.Map(taskTemplates, func(taskTemplate templatemodels.TaskTemplate) *pb.GetAllTaskTemplatesByUserResponse_TaskTemplate {
+		var mappedSubtasks = hwutil.Map(taskTemplate.SubTasks, func(subtask templatemodels.TaskTemplateSubtask) *pb.GetAllTaskTemplatesByUserResponse_TaskTemplate_SubTask {
 			return &pb.GetAllTaskTemplatesByUserResponse_TaskTemplate_SubTask{
 				Id:             subtask.ID.String(),
 				TaskTemplateId: subtask.TaskTemplateID.String(),
@@ -304,7 +284,7 @@ func (ServiceServer) GetAllTaskTemplatesByWard(ctx context.Context, req *pb.GetA
 		return nil, status.Error(codes.InvalidArgument, err.Error())
 	}
 
-	var taskTemplates []TaskTemplate
+	var taskTemplates []templatemodels.TaskTemplate
 
 	if err := db.Preload("SubTasks").Where("ward_id = ?", wardId).Find(&taskTemplates).Error; err != nil {
 		if hwgorm.IsOurFault(err) {
@@ -314,8 +294,8 @@ func (ServiceServer) GetAllTaskTemplatesByWard(ctx context.Context, req *pb.GetA
 		}
 	}
 
-	var mappedTaskTemplates = hwutil.Map(taskTemplates, func(taskTemplate TaskTemplate) *pb.GetAllTaskTemplatesByWardResponse_TaskTemplate {
-		var mappedSubtasks = hwutil.Map(taskTemplate.SubTasks, func(subtask TaskTemplateSubtask) *pb.GetAllTaskTemplatesByWardResponse_TaskTemplate_SubTask {
+	var mappedTaskTemplates = hwutil.Map(taskTemplates, func(taskTemplate templatemodels.TaskTemplate) *pb.GetAllTaskTemplatesByWardResponse_TaskTemplate {
+		var mappedSubtasks = hwutil.Map(taskTemplate.SubTasks, func(subtask templatemodels.TaskTemplateSubtask) *pb.GetAllTaskTemplatesByWardResponse_TaskTemplate_SubTask {
 			return &pb.GetAllTaskTemplatesByWardResponse_TaskTemplate_SubTask{
 				Id:             subtask.ID.String(),
 				TaskTemplateId: subtask.TaskTemplateID.String(),
